add error code
git-svn-id: http://code.taobao.org/svn/tfs@645 b722c2b5-6a69-4643-bcd1-1a75825fa998

+++ branches/tfs-dev-1.4.1/src/name_meta_server/meta_server_service.h
@@ -86,7 +86,7 @@ namespace tfs
           std::vector<MetaInfo>& v_meta_info, int32_t& cluster_id, int64_t& last_offset);
 
       int read_frag_info(const std::vector<MetaInfo>& v_meta_info,
-          const int64_t offset, const int32_t size,
+          const int64_t offset, const int64_t size,
           int32_t& cluster_id, std::vector<FragMeta>& v_out_frag_info, bool& still_have);
